DTM0: Add DTM0 service and its parameters.
The patch adds DTM0 service to the cfgen, so that
it ends up in the confc on the client and on the
server side. Also, the patch includes a change for
the cfgen that puts the DTM0 service origin into
the parameters list (origin:in-volatile,
origin:in-persistent).

NOTE: this a dirty patch, it has to be polished.